fix(rdb-inventario): mover types/const fuera de actions.ts ('use server')
Next.js (Turbopack) requiere que un archivo con 'use server' exporte SOLO
funciones async. Los `export const RDB_EMPRESA_ID` y `export type ...` que
había en actions.ts hacían que el módulo entero terminara con cero exports
visibles ("The module has no exports at all"), por lo que el import desde
nuevo/page.tsx fallaba en build aunque pasara typecheck local.

Extraído a app/rdb/inventario/levantamientos/types.ts:
- RDB_EMPRESA_ID
- ActionResult<T>, CrearLevantamientoInput, FirmarPasoInput, FirmarPasoData
- LineaParaCapturar, LineaParaRevisar

actions.ts ahora importa todo desde ./types y solo exporta las 8 funciones
async server-side. nuevo/page.tsx usa `import type` desde ./types para
CrearLevantamientoInput.

Verificado: `npm run build` ahora compila correctamente (el error pre-render
local restante es por env vars de Supabase ausentes — no aplica a Vercel).
typecheck/lint/format/test todo verde.
